Move GDALPamMultiDim and GDALPamMDArray classes to new gdal_pam_multidim.h
