+++ frmts/netcdf/netcdfmultidim.cpp
@@ -15,6 +15,7 @@
 #include <map>
 
 #include "cpl_multiproc.h"
+#include "gdal_pam_multidim.h"
 #include "gdal_rat.h"
 #include "memdataset.h"
